// src/generated/prisma/enums.ts
export const ConsultationMode = {
  IN_PERSON: 'IN_PERSON',
  VIDEO: 'VIDEO',
  PHONE: 'PHONE'
} as const

export type ConsultationMode = (typeof ConsultationMode)[keyof typeof ConsultationMode]


export const AppointmentStatus = {
  PENDING_PAYMENT: 'PENDING_PAYMENT',
  CONFIRMED: 'CONFIRMED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED_BY_PATIENT: 'CANCELLED_BY_PATIENT',
  CANCELLED_BY_DOCTOR: 'CANCELLED_BY_DOCTOR',
  NO_SHOW: 'NO_SHOW',
  EXPIRED: 'EXPIRED'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]